53 parallize md

+++ src/animl/detect.py
@@ -10,7 +10,9 @@ import pandas as pd
 from shutil import copyfile
 from PIL import Image
 from . import file_management
-
+from ultralytics import YOLO
+import multiprocessing as mp
+from numpy import vstack
 
 def process_image(im_file, detector, confidence_threshold, quiet=True,
                   image_size=None, skip_image_resize=False):
@@ -170,7 +172,49 @@ def detect_MD_batch(detector, image_file_names, checkpoint_path=None, checkpoint
     return results
 
 
-def parse_MD(results, manifest=None, out_file=None, buffer=0.02, threshold=0, checkpoint_frequency=-1):
+def process_frame(frame, threshold=0.5, buffer=0.02):
+    """
+    Takes in a single frame and returns a list of its detections
+    
+        Args:
+            - frame (dict): single md output 
+            - buffer (float): adjust bbox by percentage of img size to avoid clipping out of bounds
+            - threshold (float): parse only detections above given confidence threshold
+
+        Returns:
+            - results: list of dict, each dict represents detections on one image
+    """
+    
+    try:
+        detections = frame['detections']
+    except KeyError:
+        print('File error ', frame['file'])
+        return None
+
+    data = []
+
+    if len(detections) == 0:
+        data.append({'file': frame['file'],
+                     'max_detection_conf': frame['max_detection_conf'],
+                     'category': 0, 'conf': None, 'bbox1': None,
+                     'bbox2': None, 'bbox3': None, 'bbox4': None})
+    else:
+        for detection in detections:
+            if detection['conf'] > threshold:
+                bbox1 = min(max(detection['bbox1'], buffer), 1 - buffer)
+                bbox2 = min(max(detection['bbox2'], buffer), 1 - buffer)
+                bbox3 = min(max(detection['bbox3'], buffer), 1 - buffer)
+                bbox4 = min(max(detection['bbox4'], buffer), 1 - buffer)
+
+                data.append({'file': frame['file'],
+                             'max_detection_conf': frame['max_detection_conf'],
+                             'category': detection['category'], 'conf': detection['conf'],
+                             'bbox1': bbox1, 'bbox2': bbox2, 'bbox3': bbox3, 'bbox4': bbox4})
+
+    return data
+
+
+def parse_MD(results, manifest=None, out_file=None, buffer=0.02, threshold=0, parallel=False, workers=mp.cpu_count(), checkpoint_frequency=-1):
     """
     Converts numerical output from classifier to common name species label
 
@@ -180,11 +224,14 @@ def parse_MD(results, manifest=None, out_file=None, buffer=0.02, threshold=0, ch
         - out_file (str): path to save dataframe
         - buffer (float): adjust bbox by percentage of img size to avoid clipping out of bounds
         - threshold (float): parse only detections above given confidence threshold
+        - parallel (boolean): parallelization enabled
+        - workers (int): number of threads running
         - checkpoint_frequency (int): write results to checkpoint file every N images
 
     Returns:
         - df (pd.DataFrame): formatted md outputs, one row per detection
     """
+    # load checkpoint
     if file_management.check_file(out_file):  # checkpoint comes back empty
         df = file_management.load_data(out_file)
         already_processed = set([i['file'] for i in df])
@@ -201,64 +248,42 @@ def parse_MD(results, manifest=None, out_file=None, buffer=0.02, threshold=0, ch
     if len(results) == 0:
         raise AssertionError("'results' contains no detections")
 
-    count = 0
-    # TODO: Add parallelization
-    for frame in tqdm(results):
-        # bypass checkpointed images
-        if frame['file'] in already_processed:
-            continue
-        
-        try:
-            detections = frame['detections']
-        except KeyError:
-            print('File error ', frame['file'])
-            continue
-        if len(detections) == 0:
-            data = {'file': [frame['file']],
-                    'max_detection_conf': [frame['max_detection_conf']],
-                    'category': [0], 'conf': [None], 'bbox1': [None],
-                    'bbox2': [None], 'bbox3': [None], 'bbox4': [None]}
-            df = pd.DataFrame(data) if df.empty else pd.concat([df, pd.DataFrame(data)]).reset_index(drop=True)
+    # parallel
+    if parallel:
+        pool = mp.Pool(workers)
 
-        else:
-            for detection in detections:
-                if (detection['conf'] > threshold):
-                    data = {'file': [frame['file']],
-                            'max_detection_conf': [frame['max_detection_conf']],
-                            'category': [detection['category']], 'conf': [detection['conf']],
-                            'bbox1': [detection['bbox1']], 'bbox2': [detection['bbox2']],
-                            'bbox3': [detection['bbox3']], 'bbox4': [detection['bbox4']]}
+        processed_data = list([pool.apply(process_frame, args=(frame, threshold, buffer)) for frame in tqdm(results)])
+        processed_data = [item for sublist in processed_data for item in sublist]
+        df = pd.DataFrame(processed_data, columns=['file', 'max_detection_conf', 'category', 'conf', 'bbox1', 'bbox2', 'bbox3', 'bbox4'])
+        pool.close()
+    # traditional
+    else:
+        count = 0
+        for frame in tqdm(results):
+             #bypass checkpointed images
+            if frame['file'] in already_processed:
+                continue
+            count += 1
 
-                    df = pd.DataFrame(data) if df.empty else pd.concat([df, pd.DataFrame(data)]).reset_index(drop=True)
+            data = process_frame(frame)
+            df = pd.concat([df, pd.DataFrame(data)]).reset_index(drop=True)
 
-                    # adjust boxes with 2% buffer from image edge
-                    df.loc[df["bbox1"] > (1 - buffer), "bbox1"] = (1 - buffer)
-                    df.loc[df["bbox2"] > (1 - buffer), "bbox2"] = (1 - buffer)
-                    df.loc[df["bbox3"] > (1 - buffer), "bbox3"] = (1 - buffer)
-                    df.loc[df["bbox4"] > (1 - buffer), "bbox4"] = (1 - buffer)
-
-                    df.loc[df["bbox1"] < buffer, "bbox1"] = buffer
-                    df.loc[df["bbox2"] < buffer, "bbox2"] = buffer
-                    df.loc[df["bbox3"] < buffer, "bbox3"] = buffer
-                    df.loc[df["bbox4"] < buffer, "bbox4"] = buffer
-        count += 1
-
-        if checkpoint_frequency != -1 and count % checkpoint_frequency == 0:
-            print('Writing a new checkpoint after having processed {} images since last restart'.format(count))
-
-            assert out_file is not None
-            # Move previous checkpoint to temp file
-            checkpoint_tmp_path = None
-            if os.path.isfile(out_file):
-                checkpoint_tmp_path = out_file + '_tmp'
-                copyfile(out_file, checkpoint_tmp_path)
-
-            # Write the new checkpoint
-            file_management.save_data(df, out_file, prompt=False)
-
-            # Remove the backup checkpoint if it exists
-            if checkpoint_tmp_path is not None:
-                os.remove(checkpoint_tmp_path)
+            if checkpoint_frequency != -1 and count % checkpoint_frequency == 0:
+                print('Writing a new checkpoint after having processed {} images since last restart'.format(count))
+            
+                assert out_file is not None
+                # Move previous checkpoint to temp file
+                checkpoint_tmp_path = None
+                if os.path.isfile(out_file):
+                    checkpoint_tmp_path = out_file + '_tmp'
+                    copyfile(out_file, checkpoint_tmp_path)
+            
+                # Write the new checkpoint
+                file_management.save_data(df, out_file, prompt=False)
+            
+                # Remove the backup checkpoint if it exists
+                if checkpoint_tmp_path is not None:
+                    os.remove(checkpoint_tmp_path)
 
     if isinstance(manifest, pd.DataFrame):
         df = manifest.merge(df, left_on="Frame", right_on="file")
